pyproject.toml
[tool.black]
target-version = ['py312']

[tool.isort]
profile = "black"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "funicular-up"
version = "0.1.0"
authors = [
  { name="andywar65", email="andy.war1965@gmail.com" },
]
description = "A two part Django app for syncing files from cloud to local computer"
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "Operating System :: OS Independent",
    "Framework :: Django",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.9"
dependencies = [
    "django-tree-queries>=0.19.0",
    "django>=4.2.18",
    "django-leaflet>=0.31.0",
    "django-geojson>=4.1.0",
]

[project.urls]
Repository = "https://github.com/andywar65/funicular-up"

[project.optional-dependencies]
dev = [
    "coverage",
    "environs[django]",
]
